utils/requirements.txt
absl-py==0.15.0
astunparse==1.6.3
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1702571698061/work
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.12
clang==5.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1655412516417/work
cycler==0.11.0
dataclasses==0.8
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
docopt==0.6.2
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
h5py==3.1.0
idna==3.10
imageio==2.15.0
importlib-metadata==4.8.3
ipykernel @ file:///D:/bld/ipykernel_1620913140366/work/dist/ipykernel-5.5.5-py3-none-any.whl
ipython @ file:///D:/bld/ipython_1609697945785/work
ipython_genutils @ file:///home/conda/feedstock_root/build_artifacts/ipython_genutils_1716278396992/work
jedi @ file:///D:/bld/jedi_1605054785407/work
joblib==1.1.1
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1642858610849/work
jupyter-core @ file:///D:/bld/jupyter_core_1631852848502/work
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==2.5.1
numpy==1.19.5
oauthlib==3.2.2
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1595548966091/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==8.4.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1670414775770/work
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydicom==2.3.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1672682006896/work
pykwalify==1.8.0
pynrrd==0.4.3
pyparsing==3.1.4
pyradiomics==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.1.1
pywin32==301
pyzmq @ file:///D:/bld/pyzmq_1631793604214/work
radiomics==0.1
requests==2.27.1
requests-oauthlib==2.0.0
rsa==4.9.1
ruamel.yaml==0.18.3
ruamel.yaml.clib==0.2.8
scikeras==0.4.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
SimpleITK==2.1.1.2
six==1.15.0
sklearn==0.0
slicer==0.0.8
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2020.9.3
torch==1.10.2
torchvision==0.11.3
tornado @ file:///D:/bld/tornado_1610094881553/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1631041982274/work
typing-extensions==3.7.4.3
urllib3==1.26.20
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1699959196938/work
Werkzeug==2.0.3
wincertstore==0.2
wrapt==1.12.1
xgboost==1.5.2
zipp==3.6.0
